+++ packages/contracts/src/IContentStore.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.21;
 
 interface IContentStore {
-    event NewContent(address indexed pointer, uint16 contentSize);
+    event NewContent(address indexed pointer, uint32 contentSize);
 
-    error ContentTooLarge(uint16 contentSize, uint16 maxSize);
+    error ContentTooLarge(uint32 contentSize, uint32 maxSize);
     error ContentAlreadyExists(address pointer);
     error ContentNotFound(address pointer);
     error AddContentFailed();
@@ -12,7 +12,7 @@ interface IContentStore {
 
     function pointerExists(address pointer) external view returns (bool);
 
-    function contentLength(address pointer) external view returns (uint16 size);
+    function contentLength(address pointer) external view returns (uint32 size);
 
     function pointerForContent(
         bytes memory content
